Use only one backup register for CMD_BOOT for stm32f103 based targets
This is behavior is more akin to what most STM32F103xx applications wanting to
enter the bootloader are accustomed to. Should retain backwards compatibility
with existing applications writing to both registers, the higher one just won't
be checked by the bootloader anymore.

+++ src/stm32f103/backup.h
@@ -32,7 +32,7 @@ enum BackupRegister {
     BKP10,
 };
 
-extern void backup_write(enum BackupRegister reg, uint32_t value);
-extern uint32_t backup_read(enum BackupRegister reg);
+extern void backup_write(enum BackupRegister reg, uint16_t value);
+extern uint16_t backup_read(enum BackupRegister reg);
 
 #endif
